Move jsonStringify test from function.spec.ts (promise return type)
- Moved test from packages/run-types/src/runType/function/function.spec.ts:217-232
- Fixed import paths for reflectFunction and FunctionRunType
- All tests passing
